backend/requirements.txt
# Core Django
Django==4.2.9
django-environ==0.11.2
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-filter==23.5
django-cors-headers==4.3.1

# Celery & Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Storage & Files
boto3==1.34.21
minio==7.2.3
python-magic==0.4.27
Pillow==10.2.0

# Security & Encryption
cryptography==42.0.0
pyjwt==2.8.0
argon2-cffi==23.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
tabula-py==2.9.0

# Validation & Utilities
phonenumbers==8.13.27
email-validator==2.1.0
python-slugify==8.0.1

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0

# ASGI Server
gunicorn==21.2.0
uvicorn[standard]==0.27.0
whitenoise==6.6.0

# WebSocket & Real-time Communication
python-socketio==5.10.0
python-engineio==4.9.0
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.19.0

